fix: EXPOSED-242 [PostgreSQL] Cannot change connection setting in middle of a tansaction
Move tests to PostgreSQL package.
